# openapi.json
{
  "components": {
    "schemas": {
...
      "AddIndexerIn": {
        "properties": {
          "display_name": {
            "default": "",
            "title": "Display Name",
            "type": "string"
          },
          "field_values": {
            "additionalProperties": true,
            "default": {},
            "title": "Field Values",
            "type": "object"
          },
          "schema_name": {
            "title": "Schema Name",
            "type": "string"
          }
        },
        "required": [
          "schema_name"
        ],
        "title": "AddIndexerIn",
        "type": "object"
      },
...
      "AddSeriesIn": {
        "properties": {
          "monitored": {
            "default": true,
            "title": "Monitored",
            "type": "boolean"
          },
          "quality_profile_id": {
            "title": "Quality Profile Id",
            "type": "integer"
          },
          "root_folder_path": {
            "title": "Root Folder Path",
            "type": "string"
          },
          "search_now": {
            "default": true,
            "title": "Search Now",
            "type": "boolean"
          },
          "season_folder": {
            "default": true,
            "title": "Season Folder",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tvdb_id": {
            "title": "Tvdb Id",
            "type": "integer"
          }
        },
        "required": [
          "tvdb_id",
          "title",
          "quality_profile_id",
          "root_folder_path"
        ],
        "title": "AddSeriesIn",
        "type": "object"
      },
...
      "CalendarItemOut": {
        "properties": {
          "app": {
            "enum": [
              "radarr",
              "sonarr",
              "readarr"
            ],
            "title": "App",
            "type": "string"
          },
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "extra": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra"
          },
          "has_file": {
            "default": false,
            "title": "Has File",
            "type": "boolean"
          },
          "release_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Type"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "app",
          "title"
        ],
        "title": "CalendarItemOut",
        "type": "object"
      },
      "CalendarResponse": {
        "properties": {
          "radarr": {
            "$ref": "#/components/schemas/ServiceBlock_list_CalendarItemOut__"
          },
          "readarr": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceBlock_list_CalendarItemOut__"
              },
              {
                "type": "null"
              }
            ]
          },
          "sonarr": {
            "$ref": "#/components/schemas/ServiceBlock_list_CalendarItemOut__"
          }
        },
        "required": [
          "radarr",
          "sonarr"
        ],
        "title": "CalendarResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HistoryEventOut": {
        "properties": {
          "date": {
            "title": "Date",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "date"
        ],
        "title": "HistoryEventOut",
        "type": "object"
      },
      "HistoryItemOut": {
        "description": "One release/torrent with every history event that happened to it.",
        "properties": {
          "app": {
            "enum": [
              "radarr",
              "sonarr",
              "readarr"
            ],
            "title": "App",
            "type": "string"
          },
          "book_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Id"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "events": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/HistoryEventOut"
            },
            "title": "Events",
            "type": "array"
          },
          "movie_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Movie Id"
          },
          "quality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality"
          },
          "series_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "app",
          "title",
          "date"
        ],
        "title": "HistoryItemOut",
        "type": "object"
      },
      "HistoryPageOut": {
        "properties": {
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoryItemOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "HistoryPageOut",
        "type": "object"
      },
...
      "IndexerOut": {
        "properties": {
          "enable": {
            "default": false,
            "title": "Enable",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "privacy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy"
          },
          "protocol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol"
          }
        },
        "required": [
          "id"
        ],
        "title": "IndexerOut",
        "type": "object"
      },
...
      "ServiceBlock_list_CalendarItemOut__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CalendarItemOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "stale_age_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stale Age Seconds"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ServiceBlock[list[CalendarItemOut]]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "arrdeck",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/calendar": {
      "get": {
        "operationId": "calendar_api_v1_calendar_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 14,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calendar",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/api/v1/history/all": {
      "get": {
        "description": "Merged, consolidated history across the arrs, paged 50/app.",
        "operationId": "history_all_api_v1_history_all_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryPageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "History All",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/api/v1/indexers": {
      "get": {
        "operationId": "indexers_api_v1_indexers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IndexerOut"
                  },
                  "title": "Response Indexers Api V1 Indexers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Indexers",
        "tags": [
          "indexers"
        ]
      },
      "post": {
        "operationId": "add_indexer_api_v1_indexers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddIndexerIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Indexer Api V1 Indexers Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Indexer",
        "tags": [
          "indexers"
        ]
      }
    },
...
    "/api/v1/series": {
      "post": {
        "operationId": "add_series_api_v1_series_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSeriesIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Series Api V1 Series Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Series",
        "tags": [
          "discover"
        ]
      }
    },
...
  }
}